+++ lib/src/core/action.dart
@@ -1,4 +1,4 @@
-import 'package:mobx/src/core/base_types.dart';
+import 'package:mobx/src/core/atom_derivation.dart';
 
 class Action {
   Function _fn;
