+++ lib/src/core/atom_derivation.dart
