+++ lib/src/core/computed.dart
@@ -1,6 +1,6 @@
 import 'package:mobx/mobx.dart';
 import 'package:mobx/src/core/action.dart';
-import 'package:mobx/src/core/base_types.dart';
+import 'package:mobx/src/core/atom_derivation.dart';
 
 class ComputedValue<T> extends Atom implements Derivation {
   @override
