+++ lib/src/core/observable.dart
@@ -1,4 +1,4 @@
-import 'package:mobx/src/core/base_types.dart';
+import 'package:mobx/src/core/atom_derivation.dart';
 import 'package:mobx/src/interceptable.dart';
 import 'package:mobx/src/listenable.dart';
